workers/bot_captcha_worker/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
pytesseract
python-multipart
requests
